+++ Game/MeadowRemixOptions.cs
@@ -78,6 +78,7 @@ public class RainMeadowOptions : OptionInterface
 
     public readonly Configurable<IntroRoll> PickedIntroRoll;
     private readonly Configurable<string> LobbyMusic;
+    public readonly Configurable<bool> AnniversaryCape;
 
     public enum IntroRoll
     {
@@ -186,6 +187,8 @@ public class RainMeadowOptions : OptionInterface
         DevNightskySkin = config.Bind("DevNightskySkin", false);
 
         EnableAchievementsOnline = config.Bind("EnableAchievementsOnline", false);
+        AnniversaryCape = config.Bind("AnniversaryCape", true);
+
     }
 
     public override void Initialize()
@@ -320,6 +323,9 @@ public class RainMeadowOptions : OptionInterface
                 new OpLabel(10, 420, Translate("Playtesting Gift")),
                 new OpCheckBox(WearingCape, new Vector2(10, 390f)),
 
+                new OpLabel(120, 420, Translate("Anniversary Gift")),
+                new OpCheckBox(AnniversaryCape, new Vector2(120, 390f)),
+                
                 new OpLabel(10, 370, Translate("Introroll")),
                 introroll = new OpComboBox2(PickedIntroRoll, new Vector2(10, 340f), 160f, OpResourceSelector.GetEnumNames(null, typeof(IntroRoll)).Select(li => { li.displayName = Translate(li.displayName); return li; }).ToList()) { colorEdge = Menu.MenuColorEffect.rgbWhite },
                 downpourWarning = new OpLabel(introroll.pos.x + 170, 70, Translate("Downpour DLC is not activated, vanilla intro will be used instead")),
@@ -328,7 +334,7 @@ public class RainMeadowOptions : OptionInterface
                 new OpLabel(10, 310, Translate("Lobby Music")),
                 music = new OpComboBox2(LobbyMusic, new Vector2(10, 280f), 160f, SongsItemList()) { colorEdge = Menu.MenuColorEffect.rgbWhite },
             };
-            if (!RainMeadow.IsDev(OnlineManager.mePlayer.id))
+            if (!MatchmakingManager.instances.Values.OfType<MatchmakingManager>().Any(x => x.IsDev(OnlineManager.mePlayer.id)))
             {
                 GeneralUIArrPlayerOptions.Skip(GeneralUIArrPlayerOptions.IndexOf(devOptions)).Take(3).Do(e => e.Hidden = true);
             }