+++ Game/RainMeadow.PlayerHooks.cs
@@ -196,6 +196,12 @@ public partial class RainMeadow
         {
             return false;
         }
+
+        if (isArenaMode(out var arena) && arena.voidMasterEnabled)
+        {
+            return true;
+        }
+
         return orig(self);
     }
     private void Player_ToggleCamo(On.Player.orig_ToggleCamo orig, Player self)
@@ -409,7 +415,7 @@ public partial class RainMeadow
             ILLabel label = null;
             c.GotoNext(MoveType.After, x => x.MatchBneUn(out label));
             c.Emit(OpCodes.Ldarg_0);
-            c.Emit(OpCodes.Ldloc, 3);
+            c.Emit(OpCodes.Ldloc, 4);
             c.EmitDelegate(delegate (Player self, int i)
             {
                 VoidSpawn spawn = self.room.voidSpawns[i];
@@ -1202,12 +1208,20 @@ public partial class RainMeadow
         public bool afkSleep;
         public int manualSleepDownCounter;
         public int timeSinceShelterWakeup;
+
+
+        public bool flightlocked;
+        public int capeFlyCounter;
     }
     public int afkSleepRequiredTime = 1200;
     public int manualSleepRequiredTime = 200;
 
     private void Player_Update1(On.Player.orig_Update orig, Player self, bool eu)
     {
+
+
+
+
         if (OnlineManager.lobby != null && self.objectInStomach != null)
             self.objectInStomach.pos = self.abstractCreature.pos;
         if (isStoryMode(out var gameMode) && self.abstractCreature.IsLocal())
@@ -1342,6 +1356,73 @@ public partial class RainMeadow
             { //For reasons beyond my comprehension, PlayerBlink() has a weird, possibly unintended "else" that makes specifically Spearmaster *not* close their eyes during sleepCurlUp.
                 self.Blink(2); //Vanilla only avoids this issue because sleepCounter also closes eyes directly using Blink(). We're not using sleepCounter, so Spearmaster needs this insomnia cure.
             }
+        }
+
+        if (OnlineManager.lobby != null && !(ModManager.MSC && 
+            self.SlugCatClass == MoreSlugcats.MoreSlugcatsEnums.SlugcatStatsName.Saint && 
+            (self.KarmaCap >= 9 || (self.room.game.session is ArenaGameSession && 
+                self.room.game.GetArenaGameSession.arenaSitting.gameTypeSetup.gameType == DLCSharedEnums.GameTypeID.Challenge && 
+                self.room.game.GetArenaGameSession.arenaSitting.gameTypeSetup.challengeMeta.ascended))) && 
+                OnlineManager.lobby.configurableBools.TryGetValue("MEADOW_ANNIVERSARY", out var anniversary) && anniversary)
+        {
+
+            if (self.IsLocal())
+            {
+                if (CapeManager.HasCape(OnlineManager.mePlayer.id).HasValue && !self.isNPC)
+                {
+                    var extras = playerExtras.GetOrCreateValue(self);
+                    if (!self.Consious)
+                    {
+                        extras.capeFlyCounter = 0;
+                        extras.flightlocked = true;
+                    }
+
+                    IntVector2 zero = new IntVector2(0, 0);
+                    if (self.canJump > 0 || self.canWallJump > 0 || self.canCorridorJump > 0)
+                    {
+                        extras.flightlocked = false;
+                    }
+                    else
+                    {
+                        if (self.input[0].pckp && self.input[0].jmp && !self.input[1].jmp && !extras.flightlocked)
+                        {
+                            extras.capeFlyCounter = 100;
+                            extras.flightlocked = true;
+                        }
+                    }
+
+                    if (extras.capeFlyCounter > 0)
+                    {
+                        extras.capeFlyCounter--;
+                        IntVector2 flyDir = new IntVector2(self.input[0].x, self.input[0].y);
+                        Player.BodyModeIndex[] bannedBodyModes = [
+                            Player.BodyModeIndex.CorridorClimb,
+                            Player.BodyModeIndex.ClimbingOnBeam,
+                            Player.BodyModeIndex.WallClimb,
+                            Player.BodyModeIndex.Swimming,
+                            Player.BodyModeIndex.ClimbIntoShortCut
+                        ];
+
+                        
+                        if ((flyDir.x == 0 && flyDir.y == 0) || bannedBodyModes.Contains(self.bodyMode))
+                        {
+                            extras.capeFlyCounter = 0;
+                        }
+                        else
+                        {
+                            self.animation = Player.AnimationIndex.None;
+                            self.mainBodyChunk.vel *= 0.8f;
+                            self.mainBodyChunk.vel += flyDir.ToVector2().normalized*2f;
+                            self.airInLungs = 1f;
+                            self.standing = false;
+                            self.WeightedPush(0, 1, Custom.DirVec(self.bodyChunks[0].pos, self.bodyChunks[1].pos), 1.0f);
+                            foreach (BodyChunk chunk in self.bodyChunks) chunk.vel.y += self.gravity * Mathf.Pow(((float)Mathf.Min(20, extras.capeFlyCounter))/20.0f, 1.4f);
+                        }
+                        
+                    }
+                }
+            }
+          
         }
     }